update invitationView. Remove some block of code in JourneyInvitation
